Backtrack / revert all 2.4.x Logs development
Paulo: I do not agree with this. I'm only doing it because I have no choice.
I already have a solution which works better than it.
So I'm crediting the person who is most responsible for this as the
commit author. David.

It represents strictly and exactly the result of:
git checkout 1fcdb29 -- src/Components/Logs/
Which I was ordered to do.